requirements.txt
# Vega2.0 pinned dependencies
# Python 3.12 compatible versions, tested as of 2025-08

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Configuration & HTTP
python-dotenv==1.0.1
requests==2.32.5
httpx==0.27.2
PyYAML==6.0.2
aiofiles==24.1.0

# Logging
structlog==24.1.0

# Cryptography
cryptography==41.0.8

# Database & ORM
SQLAlchemy==2.0.35

# ML & datasets
datasets==3.0.1
transformers==4.56.0
accelerate==0.34.2

# Image processing
Pillow==10.3.0
piexif==1.1.3

# CLI & UX
rich==13.8.1
typer==0.12.5
click==8.1.7

# System Monitoring
psutil==6.1.0

# Optional rate limiting
slowapi==0.1.9

# PEFT & LoRA (optional)
peft==0.13.2
bitsandbytes==0.43.1  # optional; only if GPU + CUDA wheels available

# Web/Image Search
duckduckgo-search==6.2.10
beautifulsoup4==4.12.3
lxml==5.3.0

# Voice Processing (optional - install if needed)
# vosk==0.3.45  # For STT
# soundfile==0.12.1  # Audio processing
# pyaudio==0.2.14  # Audio I/O (may need system deps)
# pyttsx3==2.97  # Cross-platform TTS fallback

# Advanced NLP (optional)
# spacy==3.8.2  # For advanced NLP features
# torch>=2.0.0  # For local ML models
# sentence-transformers==3.0.1  # For embeddings

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
